// src/enums/email-validation-status.enum.ts
export enum EmailValidationStatus {
  PENDING = 'PENDING',
  VALID = 'VALID',
  INVALID_FORMAT = 'INVALID_FORMAT',
  NO_MX_RECORDS = 'ANY_VALID_MX_RECORDS',
  DISPOSABLE_ADDRESS = 'DISPOSABLE_ADDRESS',
  BLACKLISTED_DOMAIN = 'BLACKLISTED_DOMAIN',
  UNDELIVERABLE_ADDRESS = 'UNDELIVERABLE_ADDRESS',
  UNDECLARED_ADDRESS = 'UNDECLARED_ADDRESS',
}